requirements.txt
aiohttp==3.11.18
attrs==21.2.0
bcrypt==3.2.0
colorama==0.4.4
Jinja2==3.1.4
pillow==10.4.0
python-dateutil==2.9.0.post0
requests==2.32.3
ttkbootstrap==1.13.2
ttkthemes==3.2.2
